+++ extensions/bootstrap/CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 bootstrap extension Change Log
 2.0.0 beta under development
 ----------------------------
 
+- Bug #2361: `yii\bootstrap\NavBar::brandUrl` should default to the home URL of application (qiangxue)
 - Enh #1474: Added option to make NavBar 100% width (cebe)
 - Enh #1553: Only add navbar-default class to NavBar when no other class is specified (cebe)
 - Enh #1601: Added support for tagName and encodeLabel parameters in ButtonDropdown (omnilight)